Remove keycode per report check
The check was added without taking into account that a keycode will not
be registered in the first place if the report is full and that unregistering
a keycode simply works by setting the keycode to zero in the report. So even
if we have a unregister keycode event that is not present/pressed in the report
this will not do any harm, just add some wasted cycles that do nothing.