Use a mutex for split shared memory synchronization
...on platforms that support it, which is only ChibiOS at this point.

The problem:

The shared memory on the ChibiOS platform can be accessed concurrently
on the secondary device. This happens when the serial transaction thread,
reacts to transactions requests by the main side of the keyboard. This
can lead to race conditions and corrupted transactions.

The solution:

The access to the split shared memory is made exclusive at any given
time by guarding it with a mutex.

The previous solution using a critical section didn't really work out
because serial functions like send and receive would disable and enable
interrupts on their own, thus leaving the system with enabled interrupts
after the first transaction. The same was true for slave transaction
handlers for e.g. sync timers, those functions effectively broke the
critical section as well.