Refactor ChibiOS USB driver to buffer all endpoints
This refactoring unifies the endpoint handling for all IN and OUT
endpoints (except for EP0 and IDLE handling) for ChibiOS based
keyboards:

- All endpoints are now backed by a configurable buffer that queues
  received reports, or reports that shall be sent.
- IN endpoints are served by the `send_report` function.
- OUT endpoints are served by the `receive_report` function.

Sending is now asynchronous and doesn't block the main loop when there
is space in the sending buffer. If the sending buffer is full, a short
timeout is applied to wait for new space in the buffer. If this fails a
"disconnected" endpoint is assumed and the next report will no longer
apply a timeout. If sending fails the report is discarded.

This queueing and timeout of reports makes the sending resilient to:

1. Host USB stack implementations that don't support advanced features
   like NKRO, extra keys or mouse keys.
2. Short hick-ups in the Host USB stack where an endpoint isn't polled
   for a short amount of time. e.g. W11/W10 has a resume from suspend
   quirk where the first report with a key down event is received but a
   fast subsequent key release event can only be received ~500ms-1s
   later. The queueing of these events should prevent the stuck key.