Flush incomplete endpoint buffers in dedicated tasks
Previously this would be done in the SOF callback which is fired with a
frequency of 1KHz thus wasting many cycles for the context switching
overhead and usually doing no work as the endpoints have no data to send
most of the time. Instead we now flush the endpoints in dedicated
endpoint tasks, which is a cheap check NULL pointer check when no work
is to be done.

Signed-off-by: Stefan Kerkmann <karlk90@pm.me>