Send keyboard idle reports via buffered endpoint
The buffered endpoint takes ownership of the USB endpoint and assumes
that all reports will be send via it's API. Side stepping this will
still cause the buffer logic to be called on transfer completion (aka.
usb_endpoint_in_tx_complete_cb) which in the worst case will cause a
waiting and to be send report to be freed as we have no measures to
detect this side stepping (and shouldn't allow it anyway).