Handle idle reports and get reports requests async
The USB idle rate handling for non-nkro keyboard endpoint was
implemented via a virtual timer that fired asynchronously at the host
requested idle rate. Sending was done in the callback of the virtual
timer in an ISR context. This is no longer an option as the asynchronous
endpoints take ownership of all used USB endpoints and thus assume that
all reports will be send via it's API.

This commit introduces a report storage that holds the last successfully
send USB report for all USB IN **HID** endpoints. This report storage is
modelled in a OOP fashion as the shared endpoint with its multiple
report types complicates the handling.

With this report storage we can not only fullfill the idle rate handling
but also the mandatory get report requests for all USB HID IN endpoints
in a generic fashion.

Signed-off-by: Stefan Kerkmann <karlk90@pm.me>