platforms: rp: link ram functions to ramtext section
With the 2.0.0 pico-sdk release it is possible to re-define the custom
pico-sdk linker sections, allowing us to drop the custom linker script
which introduced the `.time_critical` section in ram. This section is
for functions we do not want to execute from flash e.g. everything that
must be fast or available if the XIP flash is deactivated e.g. when
programming the flash itself.